Update patch set 1
Patch Set 1:

I don't know and that's why I asked :)

I'm not sure why projects deploy their Karaf distro today since we have the simultaneous release that releases the official one but you may be right in that they shouldn't be assuming it's only used for developer testing. Perhaps a change that we need to apply to all projects.

Patch-set: 1