Update patch set 2
Patch Set 2:

(3 comments)

Mostly nits about checkstyle, that I think we should turn on if it is not already. The one bigger question was about the SNMP call. If I read the code correctly this is a synchronous call to the node via SNMP, just wondering if we should handle the response async as opposed to waiting for the response. Not a huge deal, but a question.

Patch-set: 2
Label: Verified=0