Update patch set 1
Patch Set 1:

The model as defined in the yang assume that we are using routed RPCs to route a VLAN request (add, remove, update, etc) to a plugin that is bound to a specific node. This plugin would then update the NE which would then trigger events which would eventually update the operational model or would update the operational model directly if the NE doesn't support change triggers.

At one point we had talked about triggering the plugins via data change notification in the /config model as opposed to exposing RPCs. The decision was based on the understanding that this was the perceived design point for the MD-SAL.

One drawback of using the data change notification is that the client receives no feedback in terms of success or failure of a request unless we add that into the model ourself.

Before submitting this patch set, I think we should verify which direction we are taking in terms of affecting configuration (data change or RPC). If we are using RPC then there are questions about if it is the responsibility of the RPC (plugin) to also update the config tree (as this is where config is persisted) and to we allow changes to be invoked both through config change and RPC or is config now just a reflected (persisted) state set via RPC. And what happens at system start when the only thing we have is config state and no RPCs called.

Not sure if all the above is clear as it is a bit of a ramble. Might want to talk these issues through on a phone call.

Patch-set: 1
Reviewer: Gerrit User 1045 <1045@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0