Update patch set 1
Patch Set 1:

This patchset raises an interesting concept that I am not sure how is meant to be handled in ODL. The issues is vendor specifics. As is, the vendor specifics are part of a "generic" project, so you get both. I am not sure how ODL's TSC recommends dealing with this. I am wondering if their approach is the generic (non-vendor) stuff in ODL. The vendor specific stuff is separate and outside ODL. Or if they think it should be in the ODL source tree. I simply don't know. We might want to look for other vendor specific code in ODL to see if it exists and / or talk to the TSC.

Patch-set: 1
Reviewer: Gerrit User 1045 <1045@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0