Update patch set 2
Patch Set 2:

(7 comments)

Should we considering keeping a history of the state changes? There are several places I would be tempted to use a list so that we have a running log of the device's operational state.

I'm also assuming a few of these items will go into the config tree, but most in operational?

Patch-set: 2
Label: Verified=0