Update patch set 2
Patch Set 2:

I like the idea of tracking state changes, but I fear the ODL storage is not set up for that. In other words, if the /oper tree is in memory, then keeping the history in memory is likely not the right approach. We need more of a "log" approach which ODL doesn't really support.

I would agree that some are more "operational" than others, but for the current set I am not sure we can clearly rule out anything other than deleted from /config. For example to invoke a sync on a NE manually you might set the /confg state to synchronizing. Same to "shutdown" comms.

Patch-set: 2