Update patch set 1
Patch Set 1:

Severity: Could you elaborate on the severity of this bug?  Is this a BLOCKER such that we cannot release Beryllium without it?  Is there a workaround such that we can write a release note and fix in Beryllium SR1?
Testing: Could you also elaborate on the testing of this patch?  How extensively has this patch been tested?  Did you do any manual/local testing?  Is it covered by any unit tests or system tests?  
Impact: Does this fix impact any dependent projects?

Patch-set: 1